[SEARCH] remove unneeded request_user from to_json
