+++ front/app/webapp/models/digitization.py
@@ -328,7 +328,7 @@ class Digitization(AbstractSearchableModel):
         type(self).objects.filter(pk=self.pk.__str__()).update(json=json_data)
         return self.json
 
-    def to_json(self, reindex=True, no_img=False, request_user=None):
+    def to_json(self, reindex=True, no_img=False):
         djson = self.json or {}
         imgs = djson.get("imgs", [] if no_img else self.get_imgs(check_in_dir=True))
         return {
