+++ front/app/webapp/models/regions.py
@@ -140,7 +140,7 @@ class Regions(AbstractSearchableModel):
             return metadata
         return {}
 
-    def to_json(self, reindex=True, no_img=False, request_user=None):
+    def to_json(self, reindex=True, no_img=False):
         """
         Do not take into account no_img for regions because there is no post_save task as for Digitization
         to fill image related properties
